pyproject.toml
[project]
name = "tick-manager"
version = "0.0.1"
description = "Ingesting, transforming, and aggregating market data"
authors = [{ name = "John Stanford", email = "787382+jxstanford@users.noreply.github.com" }]
readme = "README.md"
keywords = ['python']
requires-python = ">=3.10,<4.0"
dependencies = [
    "click>=8.1.7",
    "fastapi>=0.115.0",
    "pydantic>=2.9.2",
    "pydantic-settings>=2.5.2",
    "uvicorn>=0.30.6",
    "pyyaml>=6.0.2",
]

[project.urls]
Homepage = "https://WeModulate.github.io/tick-manager/"
Repository = "https://github.com/WeModulate/tick-manager"
Documentation = "https://WeModulate.github.io/tick-manager/"

[tool.uv]
dev-dependencies = [
    "pytest>=7.2.0",
    "pre-commit>=2.20.0",
    "tox-uv>=1.11.3",
    "deptry>=0.20.0",
    "mypy>=0.991",
    "pytest-cov>=4.0.0",
    "ruff>=0.0.235",
    "mkdocs>=1.4.2",
    "mkdocs-material>=8.5.10",
    "mkdocstrings[python]>=0.19.0",
    "httpx>=0.27.2",
    "types-pyyaml>=6.0.12.20240917",
]

[tool.mypy]
files = ["tick_manager"]
disallow_untyped_defs = "True"
disallow_any_unimported = "True"
no_implicit_optional = "True"
check_untyped_defs = "True"
warn_return_any = "True"
warn_unused_ignores = "True"
show_error_codes = "True"

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
target-version = "py39"
line-length = 120
fix = true

[tool.ruff.format]
preview = true

[tool.ruff.lint]
select = [
    # flake8-2020
    "YTT",
    # flake8-bandit
    "S",
    # flake8-bugbear
    "B",
    # flake8-builtins
    "A",
    # flake8-comprehensions
    "C4",
    # flake8-debugger
    "T10",
    # flake8-simplify
    "SIM",
    # isort
    "I",
    # mccabe
    "C90",
    # pycodestyle
    "E", "W",
    # pyflakes
    "F",
    # pygrep-hooks
    "PGH",
    # pyupgrade
    "UP",
    # ruff
    "RUF",
    # tryceratops
    "TRY",
]
ignore = [
    # LineTooLong
    "E501",
    # DoNotAssignLambda
    "E731",
    # BroadException
    "TRY003",
]
per-file-ignores = { "tests/*" = ["S101"] }

[tool.coverage.report]
skip_empty = true

[tool.coverage.run]
branch = true
source = ["tick_manager"]
